remove flag_v2
