+++ api/routes/__init__.py
@@ -1,27 +1,21 @@
-from .assessment_routes import all_assessments_flagsv2_for_fund_round_id
 from .assessment_routes import all_assessments_for_fund_round_id
 from .assessment_routes import assessment_metadata_for_application_id
-from .assessment_routes import assessment_stats_flagsv2_for_fund_round_id
 from .assessment_routes import assessment_stats_for_fund_round_id
-from .assessment_routes import create_flag_v2_for_application
-from .assessment_routes import get_all_flags_v2_for_application
 from .assessment_routes import get_all_uploaded_document_theme_answers
 from .assessment_routes import get_application_data_for_export
 from .assessment_routes import get_application_json
 from .assessment_routes import get_assessor_task_list_state
 from .assessment_routes import get_banner_state
-from .assessment_routes import get_flag_v2
 from .assessment_routes import get_sub_criteria_theme_answers
-from .assessment_routes import get_team_flag_stats
 from .assessment_routes import sub_criteria
 from .assessment_routes import update_ar_status_to_completed
-from .assessment_routes import update_flag_v2_for_application
 from .comment_routes import comments_for_application_sub_criteria
 from .comment_routes import post_comments_for_application_sub_criteria
+from .flag_routes import create_flag_for_application
 from .flag_routes import get_all_flags_for_application
 from .flag_routes import get_flag
-from .flag_routes import get_latest_flag_for_application
-from .flag_routes import post_flag_for_application
+from .flag_routes import get_team_flag_stats
+from .flag_routes import update_flag_for_application
 from .progress_routes import get_progress_for_applications
 from .progress_routes import post_progress_for_applications
 from .qa_complete_routes import post_qa_complete_for_application
@@ -29,13 +23,14 @@ from .qa_complete_routes import qa_complete_record_for_application
 from .score_routes import get_score_for_application_sub_criteria
 from .score_routes import post_score_for_application_sub_criteria
 from .tag_routes import add_tag_for_fund_round
+from .tag_routes import associate_tags_with_assessment
+from .tag_routes import get_active_tags_associated_with_assessment
 from .tag_routes import get_tag
 from .tag_routes import get_tags_for_fund_round
 
 
 __all__ = [
     "all_assessments_for_fund_round_id",
-    "all_assessments_flagsv2_for_fund_round_id",
     "assessment_metadata_for_application_id",
     "get_score_for_application_sub_criteria",
     "post_score_for_application_sub_criteria",
@@ -46,25 +41,22 @@ __all__ = [
     "comments_for_application_sub_criteria",
     "post_comments_for_application_sub_criteria",
     "get_banner_state",
-    "get_latest_flag_for_application",
-    "post_flag_for_application",
     "get_progress_for_applications",
     "assessment_stats_for_fund_round_id",
-    "assessment_stats_flagsv2_for_fund_round_id",
     "post_progress_for_applications",
     "update_ar_status_to_completed",
     "get_application_json",
-    "get_all_flags_for_application",
-    "get_flag",
     "get_team_flag_stats",
-    "get_all_flags_v2_for_application",
-    "update_flag_v2_for_application",
-    "create_flag_v2_for_application",
+    "get_all_flags_for_application",
+    "update_flag_for_application",
+    "create_flag_for_application",
     "get_tags_for_fund_round",
     "add_tag_for_fund_round",
-    "get_flag_v2",
+    "get_flag",
     "get_tag",
     "post_qa_complete_for_application",
     "qa_complete_record_for_application",
     "get_application_data_for_export",
+    "get_active_tags_associated_with_assessment",
+    "associate_tags_with_assessment",
 ]
