+++ db/models/__init__.py
@@ -2,9 +2,8 @@
 from db.models.assessment_record import AssessmentRecord
 from db.models.assessment_record import TagAssociation
 from db.models.comment import Comment
-from db.models.flags import Flag
-from db.models.flags_v2 import AssessmentFlag
-from db.models.flags_v2 import FlagUpdate
+from db.models.flags import AssessmentFlag
+from db.models.flags import FlagUpdate
 from db.models.qa_complete.qa_complete import QaComplete
 from db.models.score import Score
 from db.models.tag import Tag
@@ -14,7 +13,6 @@ __all__ = [
     "AssessmentRecord",
     "Score",
     "Comment",
-    "Flag",
     "FlagUpdate",
     "TagAssociation",
     "AssessmentFlag",
