+++ db/models/flags/__init__.py
@@ -1,4 +1,6 @@
 # flake8: noqa
-from .flags import Flag
+from .assessment_flag import AssessmentFlag
+from .flag_update import FlagStatus
+from .flag_update import FlagUpdate
 
-__all__ = ["Flag"]
+__all__ = ["AssessmentFlag", "FlagUpdate", "FlagStatus"]
