--- db/models/flags_v2/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-from .assessment_flag import AssessmentFlag
-from .flag_update import FlagUpdate
-
-__all__ = ["AssessmentFlag", "FlagUpdate"]
