+++ db/queries/__init__.py
@@ -1,14 +1,9 @@
 from .assessment_records.queries import bulk_insert_application_record
 from .assessment_records.queries import delete_assessment_record
 from .assessment_records.queries import find_answer_by_key_runner
-from .assessment_records.queries import get_metadata_flagsv2_for_fund_round_id
 from .assessment_records.queries import get_metadata_for_fund_round_id
 from .comments.queries import create_comment_for_application_sub_crit
 from .comments.queries import get_comments_for_application_sub_crit
-from .flags.queries import create_flag_for_application
-from .flags.queries import find_qa_complete_flag_for_applications
-from .flags.queries import retrieve_flag_for_application
-from .flags.queries import retrieve_flags_for_applications
 from .progress.queries import get_progress_for_app
 from .scores.queries import create_score_for_app_sub_crit
 from .scores.queries import get_scores_for_app_sub_crit
@@ -19,17 +14,12 @@ __all__ = [
     "select_tags_for_fund_round",
     "insert_tags",
     "get_metadata_for_fund_round_id",
-    "get_metadata_flagsv2_for_fund_round_id",
     "bulk_insert_application_record",
     "find_answer_by_key_runner",
     "get_scores_for_app_sub_crit",
     "create_score_for_app_sub_crit",
     "get_comments_for_application_sub_crit",
     "create_comment_for_application_sub_crit",
-    "retrieve_flag_for_application",
-    "retrieve_flags_for_applications",
-    "create_flag_for_application",
     "get_progress_for_app",
-    "find_qa_complete_flag_for_applications",
     "delete_assessment_record",
 ]
