+++ db/queries/flags/queries.py
@@ -1,16 +1,24 @@
-"""Queries which are performed on the `flags` table.
-
-Joins allowed.
-"""
 from typing import Dict
-from typing import Iterable
 
 from db import db
-from db.models import Flag
-from db.models.flags.enums import FlagType
-from db.models.flags_v2 import AssessmentFlag
-from db.schemas.schemas import FlagMetadata
-from sqlalchemy import and_
+from db.models.flags.assessment_flag import AssessmentFlag
+from db.models.flags.flag_update import FlagStatus
+from db.models.flags.flag_update import FlagUpdate
+from sqlalchemy import select
+
+
+def get_flags_for_application(application_id):
+    stmt = select(AssessmentFlag).where(
+        AssessmentFlag.application_id == application_id
+    )
+    results = db.session.scalars(stmt).all()
+    return results
+
+
+def get_flag_by_id(flag_id):
+    stmt = select(AssessmentFlag).where(AssessmentFlag.id == flag_id)
+    results = db.session.scalars(stmt).all()
+    return results
 
 
 def create_flag_for_application(
@@ -18,145 +26,52 @@ def create_flag_for_application(
     sections_to_flag: str,
     application_id: str,
     user_id: str,
-    flag_type: FlagType,
+    status: FlagStatus,
+    allocation: str,
 ) -> Dict:
-    flag = Flag(
+    flag_update = FlagUpdate(
         justification=justification,
-        sections_to_flag=sections_to_flag,
-        application_id=application_id,
         user_id=user_id,
-        flag_type=flag_type,
+        status=status,
+        allocation=allocation,
     )
-    db.session.add(flag)
+    assessment_flag = AssessmentFlag(
+        application_id=application_id,
+        sections_to_flag=sections_to_flag,
+        latest_allocation=allocation,
+        latest_status=status,
+        updates=[flag_update],
+    )
+    db.session.add(assessment_flag)
     db.session.commit()
-
-    metadata_serialiser = FlagMetadata()
-    flag_metadata = metadata_serialiser.dump(flag)
-
-    return flag_metadata
+    return assessment_flag
 
 
-def retrieve_flag(flag_id: str) -> dict:
-    flag = Flag.query.filter(Flag.id == flag_id).first()
-    metadata_serialiser = FlagMetadata()
-    flag_metadata = metadata_serialiser.dump(flag)
+def add_update_to_assessment_flag(
+    justification: str,
+    user_id: str,
+    status: FlagStatus,
+    allocation: str,
+    assessment_flag_id: str,
+) -> Dict:
 
-    return flag_metadata
-
-
-def retrieve_flag_for_application(application_id: str) -> dict:
-    flag = (
-        Flag.query.filter(Flag.application_id == application_id)
-        .order_by(Flag.date_created.desc())
-        .first()
+    stmt = select(AssessmentFlag).where(
+        AssessmentFlag.id == assessment_flag_id
     )
 
-    metadata_serialiser = FlagMetadata()
-    flag_metadata = metadata_serialiser.dump(flag)
+    assessment_flag = db.session.scalars(stmt).one()
 
-    return flag_metadata
-
-
-def retrieve_all_flags_for_application(application_id: str) -> list[dict]:
-    flags = (
-        Flag.query.filter(Flag.application_id == application_id)
-        .order_by(Flag.date_created.desc())
-        .all()
+    flag_update = FlagUpdate(
+        justification=justification,
+        user_id=user_id,
+        status=status,
+        allocation=allocation,
+        assessment_flag_id=assessment_flag_id,
     )
-    metadata_serialiser = FlagMetadata()
-    flag_metadata_list = [metadata_serialiser.dump(flag) for flag in flags]
-    return flag_metadata_list
+    assessment_flag.updates.append(flag_update)
+    assessment_flag.latest_allocation = allocation
+    assessment_flag.latest_status = status
 
-
-def find_qa_complete_flags(application_ids: Iterable[str]) -> dict[str, bool]:
-    flags = Flag.query.filter(
-        Flag.application_id.in_(application_ids),
-        Flag.flag_type == "QA_COMPLETED",
-    ).all()
-
-    qa_completed_application_ids = {flag.application_id for flag in flags}
-    return {
-        app_id: app_id in qa_completed_application_ids
-        for app_id in application_ids
-    }
-
-
-def find_qa_complete_flagsv2(
-    application_ids: Iterable[str],
-) -> dict[str, bool]:
-    flags = AssessmentFlag.query.filter(
-        AssessmentFlag.application_id.in_(application_ids),
-        AssessmentFlag.latest_status == "QA_COMPLETED",
-    ).all()
-
-    qa_completed_application_ids = {flag.application_id for flag in flags}
-    return {
-        app_id: app_id in qa_completed_application_ids
-        for app_id in application_ids
-    }
-
-
-def find_qa_complete_flag(application_id):
-    qa_complete_flag = Flag.query.filter(
-        Flag.application_id == application_id, Flag.flag_type == "QA_COMPLETED"
-    ).first()
-
-    return {"is_qa_complete": bool(qa_complete_flag)}
-
-
-def find_qa_complete_flag_for_applications(application_ids: list[str]) -> dict:
-    flags = Flag.query.filter(
-        Flag.application_id.in_(application_ids),
-        Flag.flag_type == "QA_COMPLETED",
-    ).all()
-
-    metadata_serialiser = FlagMetadata()
-    flag_metadatas = [metadata_serialiser.dump(flag) for flag in flags]
-
-    return flag_metadatas
-
-
-def retrieve_flags_for_applications(application_ids: list[str]) -> dict:
-    flags = (
-        Flag.query.filter(Flag.application_id.in_(application_ids))
-        .order_by(Flag.date_created.desc())
-        .all()
-    )
-
-    metadata_serialiser = FlagMetadata()
-    flag_metadatas = [metadata_serialiser.dump(flag) for flag in flags]
-
-    return flag_metadatas
-
-
-def get_latest_flags_for_each(flag_type: str = None) -> dict:
-
-    # Create subquery to select latest flag for applications
-    subquery = (
-        db.session.query(
-            Flag.application_id,
-            db.func.max(Flag.date_created).label("latest_date"),
-        )
-        .group_by(Flag.application_id)
-        .subquery()
-    )
-
-    # JOIN with main query of flags
-    query = db.session.query(Flag).join(
-        subquery,
-        and_(
-            Flag.application_id == subquery.c.application_id,
-            Flag.date_created == subquery.c.latest_date,
-        ),
-    )
-
-    # Add filter of flag_type if provided
-    if flag_type:
-        query = query.where(Flag.flag_type == flag_type)
-
-    flags = db.session.execute(query).scalars().all()
-
-    metadata_serialiser = FlagMetadata()
-    flag_metadatas = [metadata_serialiser.dump(flag) for flag in flags]
-
-    return flag_metadatas
+    db.session.add(assessment_flag)
+    db.session.commit()
+    return assessment_flag
